Move the general logic from tune_cystein_bridges to a separate module
